+++ automerge-backend/src/backend.rs
@@ -5,9 +5,9 @@ use crate::actor_map::ActorMap;
 use crate::op_set::OpSet;
 use crate::ordered_set::{OrdDelta, OrderedSet};
 use crate::patch::{Diff, Patch, PendingDiff};
-use crate::protocol::{ObjectID, OpType, Operation, ReqOpType, UndoOperation, OpID};
+use crate::protocol::{ObjectID, OpType, Operation, ReqOpType, UndoOperation};
 use crate::time;
-use automerge_protocol::{ActorID, ChangeHash, ObjType};
+use automerge_protocol::{ActorID, ChangeHash, ObjType, OpID};
 use crate::{Change, ChangeRequest, ChangeRequestType};
 use std::borrow::BorrowMut;
 use std::cmp::max;
