+++ automerge-backend/src/protocol.rs
@@ -9,9 +9,6 @@
 //!
 
 use serde::{Deserialize, Serialize};
-use std::cmp::{Ordering, PartialOrd};
-use std::convert::TryFrom;
-use std::fmt;
 use std::hash::Hash;
 use std::str::FromStr;
 
@@ -20,7 +17,7 @@ use crate::error::AutomergeError;
 use crate::helper;
 use crate::op_handle::OpHandle;
 use crate::ordered_set::OrderedSet;
-use automerge_protocol::{ActorID, ChangeHash, ObjType};
+use automerge_protocol::{ActorID, ChangeHash, ObjType, OpID};
 
 
 #[derive(Eq, PartialEq, Debug, Hash, Clone)]
@@ -49,40 +46,6 @@ impl FromStr for ObjectID {
     }
 }
 
-#[derive(Eq, PartialEq, Hash, Clone)]
-pub struct OpID(pub u64, pub String);
-
-impl Ord for OpID {
-    fn cmp(&self, other: &Self) -> Ordering {
-        if self.0 != other.0 {
-            self.0.cmp(&other.0)
-        } else {
-            self.1.cmp(&other.1)
-        }
-    }
-}
-
-impl fmt::Debug for OpID {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.to_string().as_str())
-    }
-}
-
-impl PartialOrd for OpID {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl OpID {
-    pub fn new(seq: u64, actor: &ActorID) -> OpID {
-        OpID(seq, actor.0.clone())
-    }
-
-    pub fn counter(&self) -> u64 {
-        self.0
-    }
-}
 
 impl From<OpID> for ObjectID {
     fn from(id: OpID) -> Self {
@@ -105,42 +68,6 @@ impl From<OpID> for Key {
 impl From<&OpID> for Key {
     fn from(id: &OpID) -> Self {
         Key::Seq(ElementID::ID(id.clone()))
-    }
-}
-
-impl FromStr for OpID {
-    type Err = AutomergeError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut i = s.split('@');
-        match (i.next(), i.next(), i.next()) {
-            (Some(seq_str), Some(actor_str), None) => seq_str
-                .parse()
-                .map(|seq| OpID(seq, actor_str.to_string()))
-                .map_err(|_| AutomergeError::InvalidOpID(s.to_string())),
-            _ => Err(AutomergeError::InvalidOpID(s.to_string())),
-        }
-    }
-}
-
-impl TryFrom<&str> for OpID {
-    type Error = AutomergeError;
-    fn try_from(s: &str) -> Result<Self, AutomergeError> {
-        OpID::from_str(s)
-    }
-}
-
-impl fmt::Display for OpID {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            OpID(seq, actor) => write!(f, "{}@{}", seq, actor),
-        }
-    }
-}
-
-impl From<&OpID> for String {
-    fn from(id: &OpID) -> Self {
-        format!("{}@{}", id.0, id.1)
     }
 }
