+++ automerge-protocol/src/error.rs
@@ -1,0 +1,5 @@
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+#[error("Invalid OpID: {0}")]
+pub struct InvalidOpID(pub String);
