+++ automerge-protocol/src/serde_impls/mod.rs
@@ -1,1 +1,2 @@
 mod change_hash;
+mod opid;
