+++ automerge-protocol/src/serde_impls/opid.rs
@@ -1,0 +1,24 @@
+use serde::de;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use crate::OpID;
+use std::str::FromStr;
+
+impl<'de> Deserialize<'de> for OpID {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        OpID::from_str(&s)
+            .map_err(|_| de::Error::invalid_value(de::Unexpected::Str(&s), &"A valid OpID"))
+    }
+}
+
+impl Serialize for OpID {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(self.to_string().as_str())
+    }
+}
